refactored all simluation specific functionality from the play/team/league class to database/generator
